Applied better solution's organization
