+++ Minandra.FluentCqlBuilder/Base/IQuery.cs
@@ -1,0 +1,7 @@
+﻿namespace Minandra.FluentCqlBuilder.Base
+{
+    public interface IQuery
+    {
+        string GetCqlStatement();
+    }
+}
