+++ Minandra.FluentCqlBuilder/Builders/Contracts/IAllowFiltering.cs
@@ -1,0 +1,7 @@
+﻿namespace Minandra.FluentCqlBuilder.Builders.Contracts
+{
+    public interface IAllowFiltering<out T> where T : class
+    {
+        T AllowFilter();
+    }
+}
