+++ Minandra.FluentCqlMapper/MapConfiguration/EntityMap.cs
@@ -1,0 +1,10 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Minandra.FluentCqlMapper.MapConfiguration
+{
+    class EntityMap
+    {
+    }
+}
