+++ Minandra.FluentCqlMapper/MapConfiguration/Interfaces/IEntityPropertyDefinition.cs
@@ -1,0 +1,9 @@
+﻿namespace Minandra.FluentCqlMapper.MapConfiguration.Interfaces
+{
+    public interface IEntityPropertyDefinition
+    {
+        string DbColumnName { get; set; }
+
+        bool Ignore { get; set; }
+    }
+}
