+++ Minandra.FluentCqlMapper/MapConfiguration/PropertyMap.cs
@@ -1,0 +1,10 @@
+﻿using Minandra.FluentCqlMapper.MapConfiguration.Interfaces;
+
+namespace Minandra.FluentCqlMapper.MapConfiguration
+{
+    public class PropertyMap : IEntityPropertyDefinition
+    {
+        public string DbColumnName { get => throw new System.NotImplementedException(); set => throw new System.NotImplementedException(); }
+        public bool Ignore { get => throw new System.NotImplementedException(); set => throw new System.NotImplementedException(); }
+    }
+}
