+++ Minandra.FluentCqlMapper/Minandra.FluentCqlMapper.csproj
