--- Minandra.FluentQueryBuilder/Base/IQuery.cs
@@ -1,7 +1,0 @@
-﻿namespace CqlQueryBuilder.Base
-{
-    public interface IQuery
-    {
-        string GetCqlStatement();
-    }
-}
