--- Minandra.FluentQueryBuilder/Builders/Contracts/IAllowFiltering.cs
@@ -1,7 +1,0 @@
-﻿namespace CqlQueryBuilder.Builders.Contracts
-{
-    public interface IAllowFiltering<out T> where T : class
-    {
-        T AllowFilter();
-    }
-}
