--- Minandra.FluentQueryBuilder/Builders/InsertBuilder.cs
@@ -1,11 +1,0 @@
-﻿using CqlQueryBuilder.Base;
-
-namespace CqlQueryBuilder.Builders
-{
-    public class InsertBuilder<T> : CqlStatementBase
-    {
-        public InsertBuilder(string query) : base(query)
-        {
-        }
-    }
-}
