+++ Minandra.Integration.Tests/UnitTest1.cs
@@ -1,0 +1,13 @@
+using Xunit;
+
+namespace Minandra.Integration.Tests
+{
+    public class UnitTest1
+    {
+        [Fact]
+        public void Test1()
+        {
+
+        }
+    }
+}
